Always await viewer data handler when loading viewer
In the case when switching between no-viewer and viewer, the data
handler promise was being cleared and reset, so content was not being
sent to viewer too soon. But in the case of an overridden viewer url
being reloaded due to presentation change, the data handler promise was
immediately resolved and so the content was sent to viewer too soon and
endless spinner resulted.